feat: 레디스 관련 메서드 추가
