--- core/src/main/java/com/dnd/sbooky/core/RedisRepository.java
@@ -1,25 +1,0 @@
-package com.dnd.sbooky.core;
-
-import java.util.concurrent.TimeUnit;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Repository;
-
-@Repository
-@RequiredArgsConstructor
-public class RedisRepository {
-
-    private final RedisTemplate<String, Object> redisTemplate;
-
-    public void setData(String key, String value, Long expiredTime) {
-        redisTemplate.opsForValue().set(key, value, expiredTime, TimeUnit.MILLISECONDS);
-    }
-
-    public String getData(String key) {
-        return (String) redisTemplate.opsForValue().get(key);
-    }
-
-    public void delete(String key) {
-        redisTemplate.delete(key);
-    }
-}
